% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_tags}
\alias{get_tags}
\title{Get value of a tag by CSS path}
\usage{
get_tags(req, tipo)
}
\arguments{
\item{req}{object returned by a \code{httr::GET} request.}

\item{tipo}{CSS path of the item (e.g. \code{#__VIEWSTATE}).}
}
\description{
This is useful for ASPX pages to get parameters inside \code{__VIEWSTATE} and
other stuff.
}
