% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtfs.R
\docType{data}
\name{gtfs}
\alias{gtfs}
\title{GTFS-SP data}
\format{A data frame with 10 rows and 2 variables:
\describe{
   \item{arq}{file name}
   \item{data}{list column with a tibble inside each element}
}}
\source{
\url{https://www.cnj.jus.br/improbidade_adm/consultar_requerido.php}
}
\usage{
gtfs
}
\description{
GTFS-SP data download on 2017-02-17.
}
\keyword{datasets}
