% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{spgtfs_download}
\alias{spgtfs_download}
\title{Download SP-GTFS data}
\usage{
spgtfs_download(login, pwd, path = "data-raw/gtfs")
}
\arguments{
\item{login}{login.}

\item{pwd}{password.}

\item{path}{folder where gtfs.zip will be written.}
}
\description{
Download SP-GTFS data from SPTrans site.
If you don't have login and password, \href{http://www.sptrans.com.br/desenvolvedores/Default.aspx?login}{click here}.
}
