% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{spgtfs_load}
\alias{spgtfs_load}
\title{Load GTFS}
\usage{
spgtfs_load(zipfile = "data-raw/gtfs/gtfs.zip")
}
\arguments{
\item{zipfile}{zip file returned by \code{spgtfs_download}.}
}
\description{
Loads spgtfs data from zip file returned by \code{spgtfs_download}.
}
